+++ CHANGELOG.md
@@ -1,7 +1,15 @@
-## [0.1.1] - 11-05-2020.
+## [1.1.0] - 24 December 2020
+
+* Added fade transition to block modal
+* Switch between fade/slide transitions using `isSlideTransitionDefault`
+* Ability to create custom transitions using `customBuildBlockModalTransitions`
+* Added ability to disable or customize 'SafeArea'
+* Docs updated
+  
+## [0.1.1] - 11 May 2020.
 
 * Readme updated.
 
-## [0.1.0] - 11-05-2020.
+## [0.1.0] - 11 May 2020.
 
 * Initial release.